requirements.txt
pyrofork==2.3.42
TgCrypto
motor==3.3.2
pymongo==4.6.1
aiohttp==3.9.3
python-dotenv==1.0.1
